# docs/openapi.yaml
openapi: 3.0.0
info:
  title: 'Identity Link DB Users API'
  description: 'API documentation for Identity Link DB Users'
  version: '1.0'
servers:
  -
    url: /users
    description: 'Users API base path'
paths:
  /api/v1/ping:
    get:
      tags:
        - Health
      summary: 'Health check endpoint'
      description: "Returns 'pong' if the service is alive."
      operationId: f7321c7b9286ff2f1842754f388964b4
      responses:
        '200':
          description: 'Pong response'
          content:
            application/json:
              schema:
                type: string
              example: pong
  /api/v1/query:
    post:
      tags:
        - Query
      summary: 'Query any entity (User, Group, etc.)'
      operationId: 8e55723842f105588b6d12bd689d80bc
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryRequest'
      responses:
        '200':
          description: 'Query result'
          content:
            application/json:
              schema:
                properties:
                  response: { properties: { result: { type: array, items: { type: object } }, hasMore: { type: boolean } }, type: object }
                type: object
  '/api/v1/group/{id}':
    get:
      tags:
        - Group
      summary: 'Fetch group by ID'
      operationId: e2c2224b9b2f903289a17b663901df6a
      parameters:
        -
          name: id
          in: path
          description: 'UUID of the group'
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 'Group fetched successfully'
          content:
            application/json:
              schema:
                properties:
                  response: { properties: { group: { $ref: '#/components/schemas/Group' } }, type: object }
                type: object
        '404':
          description: 'Group not found'
    put:
      tags:
        - Group
      summary: 'Update an existing group'
      operationId: 42f7c9ed90b6cf967e534a20b9e73b83
      parameters:
        -
          name: id
          in: path
          description: 'UUID of the group'
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Group'
      responses:
        '200':
          description: 'Group updated successfully'
          content:
            application/json:
              schema:
                properties:
                  response: { properties: { group: { $ref: '#/components/schemas/Group' } }, type: object }
                type: object
        '400':
          description: 'Invalid input'
        '403':
          description: 'System groups cannot be updated'
        '404':
          description: 'Group not found'
    delete:
      tags:
        - Group
      summary: 'Delete a group'
      operationId: 259005a9fc55a4cd03440ec4bd7c041c
      parameters:
        -
          name: id
          in: path
          description: 'UUID of the group'
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: 'Group deleted successfully'
        '403':
          description: 'System groups cannot be deleted'
        '404':
          description: 'Group not found'
  /api/v1/group:
    post:
      tags:
        - Group
      summary: 'Create a new group'
      operationId: 5d7927e8098b67ebba6c39dad87fcd8e
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Group'
      responses:
        '201':
          description: 'Group created successfully'
          content:
            application/json:
              schema:
                properties:
                  response: { properties: { group: { $ref: '#/components/schemas/Group' } }, type: object }
                type: object
        '400':
          description: 'Invalid input'
  '/api/v1/user/{id}':
    get:
      tags:
        - User
      summary: 'Fetch a user by ID'
      operationId: 93ee4e99b3ac0dc57c079214b03b3c5a
      parameters:
        -
          name: id
          in: path
          description: 'UUID of the user'
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 'User fetched successfully'
          content:
            application/json:
              schema:
                properties:
                  response: { properties: { user: { $ref: '#/components/schemas/User' } }, type: object }
                type: object
        '404':
          description: 'User not found'
    put:
      tags:
        - User
      summary: 'Update an existing user'
      operationId: e96309ee5b8a58590cf0255194b6053a
      parameters:
        -
          name: id
          in: path
          description: 'UUID of the user'
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: 'User updated successfully'
          content:
            application/json:
              schema:
                properties:
                  response: { properties: { user: { $ref: '#/components/schemas/User' } }, type: object }
                type: object
        '403':
          description: 'System users cannot be updated'
        '404':
          description: 'User not found'
    delete:
      tags:
        - User
      summary: 'Delete a user'
      operationId: 64875aeb3b8515fff34068fd050e59ac
      parameters:
        -
          name: id
          in: path
          description: 'UUID of the user'
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: 'User deleted'
        '403':
          description: 'System users cannot be deleted'
        '404':
          description: 'User not found'
  /api/v1/user:
    post:
      tags:
        - User
      summary: 'Create a new user'
      operationId: 0b5d0d296420bb286e85ee0c51944078
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: 'User created successfully'
          content:
            application/json:
              schema:
                properties:
                  response: { properties: { user: { $ref: '#/components/schemas/User' } }, type: object }
                type: object
        '404':
          description: 'User not found'
  /api/v1/auth:
    post:
      tags:
        - User
      summary: 'Authenticate user'
      operationId: d4c334b302e014700965878de46d8abd
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthUserRequest'
      responses:
        '200':
          description: 'Authentication successful'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: 'Invalid credentials'
components:
  schemas:
    QueryRequest:
      title: 'Query Request'
      description: 'Request body for generic entity querying'
      required:
        - type
      properties:
        type:
          description: 'The entity type being queried (e.g., "User", "Group")'
          type: string
          example: User
        alias:
          description: 'Alias used in the query builder'
          type: string
          default: t
          example: t
        query:
          description: 'DQL-compatible where clause, e.g., "t.name = :name"'
          type: string
          example: 'g.name = :group AND t.username = :username'
          nullable: true
        joins:
          description: 'Joins to perform (DQL-style)'
          type: object
          example:
            g: t.groups
          nullable: true
          additionalProperties: true
        orderBy:
          description: 'Ordering of results'
          type: object
          example:
            t.username: ASC
          nullable: true
          additionalProperties: true
        parameters:
          description: 'Parameters used in query, used with query string placeholders'
          type: object
          example:
            username: test
            group: administrator
          nullable: true
          additionalProperties: true
        limit:
          description: 'Maximum number of results to return'
          type: integer
          format: int32
          example: 10
        offset:
          description: 'Number of records to skip'
          type: integer
          format: int32
          example: 0
      type: object
    AuthUserRequest:
      required:
        - username
        - password
      properties:
        username:
          description: 'Username for authentication'
          type: string
          maxLength: 200
          example: john_doe
        password:
          description: 'Password for authentication'
          type: string
          maxLength: 200
          example: myS3cretPass
        grantType:
          description: 'OAuth2 grant type'
          type: string
          enum:
            - client_credentials
            - password
            - authorization_code
            - refresh_token
            - implicit
          example: password
          nullable: true
      type: object
    Group:
      title: Group
      description: 'Group entity schema used for both input and output'
      required:
        - name
      properties:
        id:
          description: 'UUID of the group'
          type: string
          format: uuid
          example: f1e2d3c4-b5a6-7890-1234-abcdef987654
        name:
          description: 'Name of the group'
          type: string
          maxLength: 100
          example: managers
        isSystem:
          description: 'Whether this group is protected from API update and deletion'
          type: boolean
          example: false
      type: object
    User:
      properties:
        id:
          description: 'User UUID'
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        username:
          description: 'Unique username'
          type: string
          example: john_doe
        password:
          description: 'User password'
          type: string
          example: superSecret123
        firstName:
          description: 'User first name'
          type: string
          example: John
        lastName:
          description: 'User last name'
          type: string
          example: Doe
        email:
          description: 'User email address'
          type: string
          format: email
          example: john.doe@example.com
        groups:
          description: 'User groups (array of UUIDs)'
          type: array
          items:
            type: string
            format: uuid
            example: 550e8400-e29b-41d4-a716-446655440000
        grantTypes:
          description: 'Allowed OAuth2 grant types'
          type: array
          items:
            type: string
            enum:
              - client_credentials
              - password
              - authorization_code
              - refresh_token
              - implicit
            example: password
        isSystem:
          description: 'Whether this user is protected from API update and deletion'
          type: boolean
          example: false
      type: object
  securitySchemes:
    bearerAuth:
      type: http
      description: 'JWT Bearer token authorization'
      bearerFormat: JWT
      scheme: bearer
security:
  -
    bearerAuth: []
tags:
  -
    name: Group
    description: Group
  -
    name: User
    description: User
  -
    name: Health
    description: Health
  -
    name: Query
    description: Query